# Migrate fake_quant_slice to TensorIterator