--- a/aten/src/ATen/native/quantized/cuda/fake_quantize_core.cu
+++ b/aten/src/ATen/native/quantized/cuda/fake_quantize_core.cu
@@ -2,6 +2,8 @@
 #include <ATen/NativeFunctions.h>
 #include <ATen/cuda/CUDAApplyUtils.cuh>
 #include <ATen/native/quantized/fake_quant_affine.h>
+#include <ATen/native/TensorIterator.h>
+#include <ATen/native/cuda/Loops.cuh>
 #include <cmath>
 
 /* Fake quantize a tensor, common block for per-channel & per-tensor fake quant
@@ -25,17 +27,22 @@ void fake_quantize_slice_kernel_cuda(
     int64_t quant_min,
     int64_t quant_max) {
   float inv_scale = 1.0f / scale;
-  at::cuda::CUDA_tensor_apply2<float, float>(
-      input, output, [=] __device__(const float& input_val, float& result_val) {
-        result_val = (fminf(
-                          quant_max,
-                          fmaxf(
-                              quant_min,
-                              static_cast<int64_t>(std::nearbyint(
-                                  input_val * inv_scale + zero_point)))) -
-                      zero_point) *
-            scale;
-      });
+  auto iter = TensorIterator();
+  iter.dont_compute_common_dtype();
+  iter.add_output(output);
+  iter.add_input(input);
+  iter.build();
+  gpu_kernel(iter,
+    [=] GPU_LAMBDA (float input_val) -> float {
+      return (fminf(
+                quant_max,
+                fmaxf(
+                    quant_min,
+                    static_cast<int64_t>(std::nearbyint(
+                        input_val * inv_scale + zero_point)))) -
+            zero_point) *
+          scale;
+    });
 }
 
 void fake_quantize_grad_slice_kernel_cuda(
@@ -47,14 +54,17 @@ void fake_quantize_grad_slice_kernel_cuda(
     int64_t quant_min,
     int64_t quant_max) {
   float inv_scale = 1.0f / scale;
-  at::cuda::CUDA_tensor_apply3<float, float, float>(
-      output_grad,
-      input,
-      input_grad,
-      [=] __device__(const float& dy, const float& x, float& dx) {
-        int64_t Xq = std::nearbyint(x * inv_scale + zero_point);
-        dx = (Xq >= quant_min && Xq <= quant_max) * dy;
-      });
+  auto iter = TensorIterator();
+  iter.dont_compute_common_dtype();
+  iter.add_output(input_grad);
+  iter.add_input(output_grad);
+  iter.add_input(input);
+  iter.build();
+  gpu_kernel(iter,
+    [=] GPU_LAMBDA (float dy, float x) -> float {
+      int64_t Xq = std::nearbyint(x * inv_scale + zero_point);
+      return (Xq >= quant_min && Xq <= quant_max) * dy;
+    });
 }
 
 REGISTER_DISPATCH(fake_quant_slice_stub, &fake_quantize_slice_kernel_cuda);