@ngimel 2020-02-25
Cool! Is per-channel quantization coming? cc @raghuramank100 @supriyar